requirements.txt
deepgram-sdk
python-dotenv
flask
flask-cors
opencv-python-headless
numpy
pandas
mediapipe
xgboost
transformers
torch
librosa